Fix error requesting network info
The script added the speedtest CLI license only at the time `Get-SpeedTestResults` was called and did not recognize the license the first time it ran on a new system, which returned an error.

The code installing the CLI license was moved out of the function to install when the script is run.