Patch no-internet-download-loop bug
This bug was found while testing the toast message feature during download. Coincidentally, the internet was disabled as a result of prior tests and a download could not be initiated.

The result of this bug is a loop at `if ($null -ne $Script:QB_VERSION)` when it checks if a version has already been selected, which would always returns yes after failing to download.

Bug fixed by making the script check for an internet connection before every download and return an error if failed.