Enforce security when enabling TLS 1.2
This commit is a precautionary measure against a possible future security risk with the current method of enabling TLS 1.2. This risk was pointed out by u/purplemonkeymad in a post on r/PowerShell titled "PSA: Enabling TLS1.2 and you", where he stated that the current method would affect the status of other protocols.

Link: https://www.reddit.com/r/PowerShell/comments/ozr6ye/psa_enabling_tls12_and_you/