test: remove custom licensing function
It's not totally necessary to have the user enter their custom (purchased) license information in the script when they can enter it themselves on the license section of the installer.

The code for `Get-UserOwnLicense` is only commented out from `PosLicensing.ps1` but will be removed before release if it is no longer needed.