Add download button and update cautionary text etc.
Initially, I only had a hyperlinked "Download Now!" element at the end of the header text right below the qbactivator logo; but apparently that was not enough since––who I assume to be––non-github users still had trouble downloading qbactivator the right way. They insisted on ignoring the link after it was highlighted for them on Telegram when they came seeking assistance, and instead continued to download the project repository which did not include the script.

This happened enough that I decided to create a large, green, glowing button that they couldn't miss.

I'm not mad; I'm rather perplexed as to how they could miss such a clear indication.

Also, cautionary text in README.md was updated again to be simpler for people who may not have understood clearly. Other grammar and semantics were updated as well for better understanding.

Information related to a specific issue commonly found with older QB software and operating systems which involves Microsoft Visual C++ packages was provided as well.